new file:   .phpunit.result.cache 	modified:   app/Console/Kernel.php 	modified:   app/Customs/Services/EmailVerificationService.php 	new file:   app/Customs/Services/ResetPasswordService.php 	modified:   app/Http/Controllers/AuthController.php 	deleted:    app/Http/Controllers/CartController.php 	modified:   app/Http/Controllers/CategoryController.php 	modified:   app/Http/Controllers/LocationController.php 	modified:   app/Http/Controllers/OrderController.php 	modified:   app/Http/Controllers/PaymentController.php 	modified:   app/Http/Controllers/ProductController.php 	new file:   app/Http/Controllers/RatingsController.php 	deleted:    app/Models/CartItem.php 	modified:   app/Models/Category.php 	modified:   app/Models/Payment.php 	new file:   app/Models/Ratings.php 	modified:   app/Models/User.php 	modified:   app/Models/location.php 	modified:   app/Models/product.php 	modified:   app/Notifications/EmailVerificationNotification.php 	new file:   app/Notifications/ResetPasswordNotification.php 	renamed:    database/migrations/2024_05_09_121030_create_cart_items_table.php -> database/migrations/2025_06_17_190052_create_ratings_table.php 	modified:   routes/api.php

+++ routes/api.php
@@ -2,12 +2,12 @@
 
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoryController;
-use App\Http\Controllers\CartController;
 use App\Http\Controllers\ProductController;
 use App\Http\Controllers\LocationController;
 use App\Http\Controllers\OrderController;
 use App\Http\Controllers\PaymentController;
 
+use App\Http\Controllers\RatingsController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -33,16 +33,23 @@ Route::group([
     Route::any('auth/verify-user-email',[AuthController::class,'verifyUserEmail']);
     Route::post('auth/resend-email-verification-code',[AuthController::class,'resendVerificationEmailCode']);
     Route::post('auth/logout', [AuthController::class, 'logout'])->middleware(['auth']);
-    Route::get('auth/user-profile', [AuthController::class, 'userProfile'])->middleware(['auth', 'verified']);
+    Route::get('user-profile', action: [AuthController::class, 'userProfile'])->middleware(['auth', 'verified']);
+    Route::post('user-profile/update', action: [AuthController::class, 'profileUpdate'])->middleware(['auth', 'verified']);
+
+
+    Route::post('auth/forgot-password', [AuthController::class, 'sendResetCode']);
+    Route::post('auth/resend-reset-code', [AuthController::class, 'resendResetCode']);
+    Route::post('auth/reset-password', [AuthController::class, 'resetPassword']);
+
 });
 
 
 Route::controller(CategoryController::class)->group(function () {
     Route::get('/category/index', 'index');
     Route::get('/category/show/{id}', 'show_products');
-    Route::post('/category/update/{id}', 'update')->middleware('is_admin');
-    Route::post('/category/create', 'store')->middleware('is_admin');
-    Route::delete('/category/delete/{id}', 'destroy')->middleware('is_admin');
+    Route::post('/admin/category/update/{id}', 'update')->middleware('is_admin');
+    Route::post('/admin/category/create', 'store')->middleware('is_admin');
+    Route::delete('/admin/category/delete/{id}', 'destroy')->middleware('is_admin');
 });
 Route::controller(LocationController::class)->group(function () {
     Route::get('/location/get_user_locations','get_user_locations')->middleware(['auth', 'verified']);
@@ -53,33 +60,31 @@ Route::controller(LocationController::class)->group(function () {
 Route::controller(ProductController::class)->group(function () {
     Route::get('/product/index', 'index');
     Route::get('/product/trendy', 'trendy_products');
+    Route::get('/product/offers', 'offers_products');
     Route::get('/product/show/{id}', 'show');
     Route::get('/product/search/{key}', 'search');
-    Route::post('/product/check_availability', 'check_availability')->middleware(['auth','verified']);
-    Route::post('/product/update/{id}', 'update')->middleware('is_admin');
-    Route::post('/product/create', 'store')->middleware('is_admin');
-    Route::delete('/product/delete/{id}', 'destroy')->middleware('is_admin');
+    Route::post('/admin/product/update/{id}', 'update')->middleware('is_admin');
+    Route::post('/admin/product/create', 'store')->middleware('is_admin');
+    Route::delete('/admin/product/delete/{id}', 'destroy')->middleware('is_admin');
 });
 
 
 
-Route::controller(CartController::class)->group(function () {
-    Route::get('/cart/get_user_cart','get_user_cart')->middleware(['auth', 'verified']);
-    Route::post('/cart/add_to_cart', 'add_to_cart')->middleware(['auth', 'verified']);
-    Route::post('/cart/remove_from_cart', 'remove_from_cart')->middleware(['auth', 'verified']);
-    Route::post('/cart/update', 'update')->middleware(['auth', 'verified']);
-});
+Route::middleware('auth:api')->post('user-profile/change-password', [AuthController::class, 'change_password']);
+Route::delete('/user-profile/delete-user', [AuthController::class, 'deleteUser']);
+
+Route::post('/rating/create', [RatingsController::class,'store'])->middleware(['auth', 'verified']);
 
 
 
 
 
 Route::controller(OrderController::class)->group(function () {
-    Route::get('/order/index', 'index')->middleware('is_admin');
     Route::get('/order/show/{id}', 'show')->middleware(['auth', 'verified']);
     Route::post('/order/create', 'store')->middleware(['auth', 'verified']);
     Route::get('/order/get_user_orders', 'get_user_orders')->middleware(['auth', 'verified']);
-    Route::post('/order/change_order_status/{id}', 'change_order_status')->middleware('is_admin');
+    Route::get('/admin/order/index', 'index')->middleware('is_admin');
+    Route::post('/admin/order/change_order_status/{id}', 'change_order_status')->middleware('is_admin');
     Route::delete('/order/delete/{id}', 'destroy')->middleware(['auth', 'verified']);
 
 });
